+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Release Notes for Blitz
 
-## 2.1.5 - 2019-07-XX
+## 2.1.5 - 2019-07-21
 ### Added
-- Added the ability to dynamically get the CSRF param and a CSRF token with the `craft.blitz.csrfParam` and `craft.blitz.csrfToken` template tags and the `CsrfController::actionParam` and `CsrfController::actionToken` controller actions.
+- Added the ability to dynamically fetch the CSRF param and a CSRF token with the `craft.blitz.csrfParam` and `craft.blitz.csrfToken` template tags and the `CsrfController::actionParam` and `CsrfController::actionToken` controller actions.
 - Added the `blitz/cache/generate-expiry-dates` console command that generates entry expiry dates and stores them to enable refreshing expired cache (this generally happens automatically).
 
 ### Fixed